Added Version sharing. (#465)
Versions can have a sharing value setup to control which project(s) they
will appear on:

* none (default) - version will only appear on it's project
* hierarchy - version will appear to parent/child projects also
* system - version will appear to any project

When a version is displayed and/or linked for unauthorized users,
it is masked to prevent information exposure.  The hidden versions will
still appear as select options though, so they are not cleared on accident.

If a Journal is entered that changes the Version to one that isn't accessible
by all of the email recipients, then the Version text is hidden.  This is
because only one email is sent to everyone.

When a Project is moved or a Version is updated, all the issues are checked
to make sure the version is still reachable according to it's sharing state.